third_party/WebKit/LayoutTests/platform/mac/external/wpt/web-animations/timing-model/timelines/document-timelines-expected.txt

This is a testharness.js-based test.
FAIL Document timelines report current time relative to navigationStart assert_equals: The current time matches requestAnimationFrame time expected 40.46200000004774 but got 44.00433333330511
Harness: the test ran to completion.
